requirements_dev.txt
torch
torchvision
cython
matplotlib
numpy
scipy
# opencv
pyyaml
packaging
pycocotools
tensorboardx
h5py
opencv-python

requests
tqdm
streamlit
pillow # for PIL

datasets
transformers 
sentence_transformers
langchain
python-box
ensure